@@ -8,7 +8,7 @@
     "test": "cross-env FORCE_COLOR=1 jest --coverage",
     "test:watch": "jest --watch",
     "posttest": "npm run lint",
-    "lint": "eslint --ignore-path ./.gitignore bin src test",
+    "lint": "eslint --ignore-path ./.gitignore src test",
     "build": "cross-env BABEL_ENV=production rollup -c",
     "prepublishOnly": "npm run build && node ./bin/updtr --help",
     "release": "standard-version",
@@ -45,11 +45,9 @@
     "babel-preset-latest": "^6.24.1",
     "clone": "^2.1.1",
     "cross-env": "^5.0.1",
-    "eslint": "^3.19.0",
-    "eslint-config-peerigon": "^10.1.1",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jest": "^20.0.3",
-    "eslint-plugin-jsdoc": "^3.1.0",
+    "eslint": "^5.6.1",
+    "eslint-config-peerigon": "^15.0.2",
+    "eslint-plugin-jest": "^21.24.0",
     "jest": "^20.0.4",
     "rimraf": "^2.6.1",
     "rollup": "^0.43.0",
