@@ -4,21 +4,21 @@ import Error from "es6-error";
 
 export class PackageJsonNoAccessError extends Error {
     constructor(dir) {
-        super(`Cannot access package.json in ${ dir }`);
+        super(`Cannot access package.json in ${dir}`);
     }
 }
 
 export class RequiredOptionMissingError extends Error {
     constructor(optionName, optionValue) {
         super(
-            `Required option ${ optionName } is missing. Instead received ${ optionValue }`
+            `Required option ${optionName} is missing. Instead received ${optionValue}`
         );
     }
 }
 
 export class OptionValueNotSupportedError extends Error {
     constructor(optionName, unsupportedValue) {
-        super(`Unsupported value ${ unsupportedValue } for option ${ optionName }`);
+        super(`Unsupported value ${unsupportedValue} for option ${optionName}`);
     }
 }
 
