@@ -1,5 +1,5 @@
 function installFn(baseCmd) {
-    return ({ registry, modules } = {}) =>
+    return ({registry, modules} = {}) =>
         [baseCmd, stringifyRegistry(registry), stringifyModules(modules)].join(
             ""
         );
@@ -11,23 +11,23 @@ function stringifyModules(modules) {
               modules
                   // We need to wrap this in double-quotes because some semver
                   // characters like the caret symbol are reserved characters on Windows.
-                  .map(({ name, version }) => `"${ name }@${ version }"`)
+                  .map(({name, version}) => `"${name}@${version}"`)
                   .join(" ") :
         "";
 }
 
 function stringifyRegistry(registry) {
-    return registry === undefined ? "" : ` --registry "${ registry }"`;
+    return registry === undefined ? "" : ` --registry "${registry}"`;
 }
 
-export default {
+const cmds = {
     npm: {
         outdated: () => "npm outdated --json --depth=0",
         installMissing: installFn("npm install"),
         install: installFn("npm install"),
         // remove: ({ name }) => ["npm remove ", name].join(""),
         test: () => "npm test",
-        list: ({ modules } = {}) => [
+        list: ({modules} = {}) => [
             "npm ls --json --depth=0",
             Array.isArray(modules) === true ? " " + modules.join(" ") : "",
         ].join(""),
@@ -40,9 +40,11 @@ export default {
         install: installFn("yarn add"),
         // remove: ({ name }) => ["yarn remove ", name].join(""),
         test: () => "yarn test",
-        list: ({ modules } = {}) => [
+        list: ({modules} = {}) => [
             "yarn list --json --depth=0",
             Array.isArray(modules) === true ? " " + modules.join(" ") : "",
         ].join(""),
     },
 };
+
+export default cmds;
