@@ -2,4 +2,6 @@ import dense from "./dense";
 import none from "./none";
 
 // The first property here is the default reporter
-export default { dense, none };
+const reporters = {dense, none};
+
+export default reporters;
