@@ -1,5 +1,5 @@
 import Sequence from "./util/Sequence";
-import { isUpdateToNonBreaking } from "./util/createUpdateTask";
+import {isUpdateToNonBreaking} from "./util/createUpdateTask";
 
 function isIncompleteResult(result) {
     return result.success === true && isUpdateToNonBreaking(result) === true;
@@ -8,8 +8,7 @@ function isIncompleteResult(result) {
 async function finishIncomplete(sequence, incomplete, allResults) {
     const updtr = sequence.updtr;
     const modulesToCheck = incomplete.map(result => result.name);
-    const listCmd = updtr.cmds.list({ modules: modulesToCheck });
-
+    const listCmd = updtr.cmds.list({modules: modulesToCheck});
     let stdout;
 
     try {
@@ -32,16 +31,16 @@ async function finishIncomplete(sequence, incomplete, allResults) {
 
                 const version = moduleVersions.find(
                     module => module.name === result.name
-            ).version;
+                ).version;
 
                 return {
                     ...result,
                     updateTo: version,
                 };
             })
-    // Remove results where no actual update did happen.
-    // These results can happen if the updateTo option was set to non-breaking
-    // and the module did not have a new version for the rollbackTo version range.
+        // Remove results where no actual update did happen.
+        // These results can happen if the updateTo option was set to non-breaking
+        // and the module did not have a new version for the rollbackTo version range.
             .filter(result => result.rollbackTo !== result.updateTo)
     );
 }
@@ -53,9 +52,9 @@ export default (async function finish(updtr, results) {
 
     if (incomplete.length > 0) {
         sequence.start();
-        sequence.emit("incomplete", { incomplete });
+        sequence.emit("incomplete", {incomplete});
         finishedResults = await finishIncomplete(sequence, incomplete, results);
-        sequence.end({ results: finishedResults });
+        sequence.end({results: finishedResults});
     }
 
     return finishedResults;
