@@ -1,5 +1,5 @@
 export default class ExecError extends Error {
-    constructor({ message, stdout = "", stderr = "", exitCode }) {
+    constructor({message, stdout = "", stderr = "", exitCode}) {
         super(message);
         this.stdout = stdout;
         this.stderr = stderr;
